tools/pyproject.toml
[project]
name = "patrickdesjardins-tools"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers",
    "scikit-learn",
    "PyYAML>=6.0",
    "requests>=2.32.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0",
]
